프론트 API 연결 프론트 아토믹 패턴 적용

+++ backend/src/main/java/jaeik/growfarm/service/comment/CommentUpdateService.java
@@ -98,9 +98,13 @@ public class CommentUpdateService {
      * @since 1.0.0
      */
     @Transactional
-    public void hardDelete(Long commentId, Comment comment) {
-        commentClosureRepository.deleteByDescendantId(commentId);
-        commentRepository.delete(comment);
+    public void commentDelete(boolean hasDescendants, Long commentId, Comment comment) {
+        if (hasDescendants) {
+            comment.softDelete();
+        } else {
+            commentClosureRepository.deleteByDescendantId(commentId);
+            commentRepository.delete(comment);
+        }
     }
